apply same undname.exe exception to WSL-CL
