+++ lib/compilers/WSL-CL.js
@@ -30,6 +30,7 @@
 
 var Compile = require('../base-compiler');
 var asm = require('../asm-cl');
+var RunCLDemangler = require('../cl-support').RunCLDemangler;
 
 function compileCl(info, env) {
     var compile = new Compile(info, env);
@@ -53,6 +54,11 @@ function compileCl(info, env) {
     compile.supportsObjdump = function () {
         return false;
     };
+
+    compile.postProcessAsm = function(result) {
+        return RunCLDemangler(this, result);
+    };
+
     compile.optionsForFilter = function (filters, outputFilename) {
         return [
             '/FAsc',
