+++ lib/compilers/Wine-CL.js
@@ -24,10 +24,7 @@
 
 var Compile = require('../base-compiler');
 var asm = require('../asm-cl');
-var utils = require('../utils');
-var fs = require('fs-extra'),
-    path = require('path'),
-    _ = require('underscore-node');
+var RunCLDemangler = require('../cl-support').RunCLDemangler;
 
 function compileCl(info, env) {
     var compile = new Compile(info, env);
@@ -51,37 +48,8 @@ function compileCl(info, env) {
         return false;
     };
 
-    compile.postProcessAsm = function (result) {
-        if (!result.okToCache) return result;
-        var demangler = this.compiler.demangler;
-        if (!demangler) return result;
-
-        if (demangler.toLowerCase().endsWith("undname.exe")) {
-            return this.newTempDir().then(function (tmpDir) {
-                var tmpfile = path.join(tmpDir, "output.s");
-                fs.writeFileSync(tmpfile, _.pluck(result.asm, 'text').join("\r\n"));
-
-                return this.exec(demangler, [tmpfile], {})
-                    .then(_.bind(function (demangleResult) {
-                    fs.unlink(tmpfile, function() {
-                        fs.remove(tmpDir, function() {});
-                    });
-
-                    var lines = utils.splitLines(demangleResult.stdout);
-                    for (var i = 0; i < result.asm.length; ++i)
-                        result.asm[i].text = lines[i];
-                    return result;
-                }, this));
-            }.bind(this));
-        } else {
-            return this.exec(demangler, [], {input: _.pluck(result.asm, 'text').join("\n")})
-                .then(_.bind(function (demangleResult) {
-                    var lines = utils.splitLines(demangleResult.stdout);
-                    for (var i = 0; i < result.asm.length; ++i)
-                        result.asm[i].text = lines[i];
-                    return result;
-                }, this));
-        }
+    compile.postProcessAsm = function(result) {
+        return RunCLDemangler(this, result);
     };
 
     compile.optionsForFilter = function (filters, outputFilename, userOptions) {
